Adding Mindless Behavior, Laptops, and Macbooks; minor change to One Direction

+++ catscripts/Entertainment/Music/Bands/One_Direction.sh
@@ -1,26 +1,29 @@
 #!/bin/bash
 
-egrep 'One Direction' newpages.txt >> OneDirection.txt
-egrep -i 'Harry (|Edward) Style' newpages.txt >> HarryStyles.txt
-egrep -i 'horan' newpages.txt >> NiallHoran.txt
-egrep -i 'Zayn(| )Malik' newpages.txt >> ZaynMalik.txt
-egrep -i 'Louis(| William) Tomlinson' newpages.txt >> LouisTomlinson.txt
-egrep -i 'Liam(| James) Payne' newpages.txt >> LiamPayne.txt
+KEYWORDS_HARRYSTYLES="Harry(| )(|Edward)(| )Style"
+KEYWORDS_NIALLHORAN="Horan"
+KEYWORDS_ZAYNMALIK="Zayn(| )Malik"
+KEYWORDS_LOUISTOMLINSON="Louis(| )(|William)(| )Tomlinson"
+KEYWORDS_LIAMPAYNE="Liam(| )(|James)(| )Payne"
+KEYWORDS_ONEDIRECTION="One(| )Directtion|1D\b|$KEYWORDS_HARRYSTYLES|$KEYWORDS_NIALLHORAN|$KEYWORDS_ZAYNMALIK|$KEYWORDS_LOUISTOMLINSON|$KEYWORDS_LIAMPAYNE"
 
-cat HarryStyles.txt NiallHoran.txt ZaynMalik.txt LouisTomlinson.txt LiamPayne.txt >> OneDirection.txt
-
-ONEDIRECTION=`stat --print=%s OneDirection.txt`
-
-if [ $ONEDIRECTION -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="OneDirection.txt"
-  export CATNAME="One Direction"
-  $CATEGORIZE
-fi
 
-rm OneDirection.txt
-rm HarryStyles.txt
-rm NiallHoran.txt
-rm ZaynMalik.txt
-rm LouisTomlinson.txt
-rm LiamPayne.txt
+  debug_start "One Direction"
+
+  ONEDIRECTION="$(egrep -i "$KEYWORDS_ONEDIRECTION" newpages.txt)"
+
+  if [ "$ONEDIRECTION" != "" ];
+  then
+    printf "%s" "$ONEDIRECTION" > OneDirection.txt
+    export CATFILE="OneDirection.txt"
+    export CATNAME="One Direction"
+    $CATEGORIZE
+    rm OneDirection.txt
+    unset ONEDIRECTION
+  fi
+
+  debug_end "One Direction"
+
+fi
